finish up cleanup of calculate datamodel
