+++ introspection-engine/connectors/sql-introspection-connector/src/misc_helpers.rs
@@ -1,10 +1,14 @@
 use datamodel::common::names::NameNormalizer;
+use datamodel::DatabaseName::{Compound, Single};
 use datamodel::{
-    dml, Field, FieldArity, FieldType, IdInfo, IdStrategy, OnDeleteStrategy, RelationInfo, ScalarType, ScalarValue,
+    dml, Field, FieldArity, FieldType, IdInfo, IdStrategy, IndexDefinition, Model, OnDeleteStrategy, RelationInfo,
+    ScalarType, ScalarValue,
 };
 use log::debug;
 use regex::Regex;
-use sql_schema_describer::{Column, ColumnTypeFamily, ForeignKey, Index, IndexType, SqlSchema, Table};
+use sql_schema_describer::{
+    Column, ColumnArity, ColumnTypeFamily, ForeignKey, ForeignKeyAction, Index, IndexType, SqlSchema, Table,
+};
 
 //checks
 
@@ -39,7 +43,16 @@ pub(crate) fn is_prisma_join_table(table: &Table) -> bool {
         && table.indices[0].tpe == IndexType::Unique
 }
 
-pub fn create_many_to_many_field(foreign_key: &ForeignKey, relation_name: String, is_self_relation: bool) -> Field {
+pub(crate) fn is_compound_foreign_key_column(table: &Table, column: &Column) -> bool {
+    match table.foreign_keys.iter().find(|fk| fk.columns.contains(&column.name)) {
+        Some(fk) if fk.columns.len() > 1 => true,
+        _ => false,
+    }
+}
+
+//calculators
+
+pub fn calculate_many_to_many_field(foreign_key: &ForeignKey, relation_name: String, is_self_relation: bool) -> Field {
     let inflector = prisma_inflector::default();
 
     let field_type = FieldType::Relation(RelationInfo {
@@ -70,14 +83,147 @@ pub fn create_many_to_many_field(foreign_key: &ForeignKey, relation_name: String
     }
 }
 
-pub(crate) fn is_compound_foreign_key_column(table: &Table, column: &Column) -> bool {
-    match table.foreign_keys.iter().find(|fk| fk.columns.contains(&column.name)) {
-        Some(fk) if fk.columns.len() > 1 => true,
-        _ => false,
-    }
+pub(crate) fn calculate_backrelation_field(
+    schema: &SqlSchema,
+    model: &&Model,
+    relation_field: &&Field,
+    relation_info: &RelationInfo,
+) -> Field {
+    let table = schema.table_bang(model.name.as_str());
+    let fk = table.foreign_key_for_column(relation_field.name.as_str());
+    let on_delete = match fk {
+        Some(fk) if fk.on_delete_action == ForeignKeyAction::Cascade => OnDeleteStrategy::Cascade,
+        _ => OnDeleteStrategy::None,
+    };
+    let field_type = FieldType::Relation(RelationInfo {
+        name: relation_info.name.clone(),
+        to: model.name.clone(),
+        to_fields: vec![relation_field.name.clone()],
+        on_delete,
+    });
+
+    let other_is_unique = || {
+        let table = schema.table_bang(&model.name);
+
+        match &relation_field.database_name {
+            None => table.is_column_unique(relation_field.name.as_str()),
+            Some(Single(name)) => table.is_column_unique(name),
+            Some(Compound(names)) => table.indices.iter().any(|i| i.columns == *names),
+        }
+    };
+    let arity = match relation_field.arity {
+        FieldArity::Required | FieldArity::Optional if other_is_unique() => FieldArity::Optional,
+        FieldArity::Required | FieldArity::Optional => FieldArity::List,
+        FieldArity::List => FieldArity::Optional,
+    };
+    let inflector = prisma_inflector::default();
+    let name = match arity {
+        FieldArity::List => inflector.pluralize(&model.name).camel_case(), // pluralize
+        FieldArity::Optional => model.name.clone().camel_case(),
+        FieldArity::Required => model.name.clone().camel_case(),
+    };
+    let field = Field {
+        name,
+        arity,
+        field_type,
+        database_name: None,
+        default_value: None,
+        is_unique: false,
+        id_info: None,
+        documentation: None,
+        is_generated: false,
+        is_updated_at: false,
+    };
+    field
 }
 
-//calculators
+pub(crate) fn calculate_compound_field(schema: &SqlSchema, table: &Table, foreign_key: &ForeignKey) -> Field {
+    debug!("Handling compound foreign key  {:?}", foreign_key);
+    let field_type = FieldType::Relation(RelationInfo {
+        name: calculate_relation_name(schema, foreign_key, table),
+        to: foreign_key.referenced_table.clone(),
+        to_fields: foreign_key.referenced_columns.clone(),
+        on_delete: OnDeleteStrategy::None,
+    });
+    let columns: Vec<&Column> = foreign_key
+        .columns
+        .iter()
+        .map(|c| table.columns.iter().find(|tc| tc.name == *c).unwrap())
+        .collect();
+    let arity = match columns.iter().find(|c| c.is_required()).is_none() {
+        true => FieldArity::Optional,
+        false => FieldArity::Required,
+    };
+    // todo this at some point needs to be a compound value of the two columns defaults?
+    let default_value = None;
+    let is_unique = false;
+    let name = foreign_key.referenced_table.clone().camel_case();
+    let database_name = Some(Compound(columns.iter().map(|c| c.name.clone()).collect()));
+    let field = Field {
+        name,
+        arity,
+        field_type,
+        database_name,
+        default_value,
+        is_unique,
+        id_info: None,
+        documentation: None,
+        is_generated: false,
+        is_updated_at: false,
+    };
+    field
+}
+
+pub(crate) fn calculate_index(index: &Index) -> IndexDefinition {
+    debug!("Handling index  {:?}", index);
+    let tpe = match index.tpe {
+        IndexType::Unique => datamodel::dml::IndexType::Unique,
+        IndexType::Normal => datamodel::dml::IndexType::Normal,
+    };
+    let index_definition: IndexDefinition = IndexDefinition {
+        name: Some(index.name.clone()),
+        fields: index.columns.clone(),
+        tpe,
+    };
+    index_definition
+}
+
+pub(crate) fn calculate_non_compound_field(schema: &&SqlSchema, table: &&Table, column: &&Column) -> Field {
+    debug!("Handling column {:?}", column);
+    let field_type = calculate_field_type(&schema, &column, &table);
+    let arity = match column.tpe.arity {
+        ColumnArity::Required => FieldArity::Required,
+        ColumnArity::Nullable => FieldArity::Optional,
+        ColumnArity::List => FieldArity::List,
+    };
+    let id_info = calculate_id_info(&column, &table);
+    let default_value = match field_type {
+        FieldType::Relation(_) => None,
+        _ if arity == FieldArity::List => None,
+        _ => column
+            .default
+            .as_ref()
+            .and_then(|default| calculate_default(default, &column.tpe.family)),
+    };
+    let is_unique = match field_type {
+        datamodel::dml::FieldType::Relation(..) => false,
+        _ if id_info.is_some() => false,
+        _ => table.is_column_unique(&column.name),
+    };
+    let field = Field {
+        name: column.name.clone(),
+        arity,
+        field_type,
+        database_name: None,
+        default_value,
+        is_unique,
+        id_info,
+        documentation: None,
+        is_generated: false,
+        is_updated_at: false,
+    };
+    field
+}
 
 pub(crate) fn calculate_default(default: &str, tpe: &ColumnTypeFamily) -> Option<ScalarValue> {
     match tpe {
@@ -183,6 +329,37 @@ pub(crate) fn calculate_field_type(schema: &SqlSchema, column: &Column, table: &
             }
         }
     }
+}
+
+// misc
+
+pub fn deduplicate_names_of_fields_to_be_added(fields_to_be_added: &mut Vec<(String, Field)>) {
+    let mut duplicated_relation_fields = Vec::new();
+    fields_to_be_added
+        .iter()
+        .enumerate()
+        .for_each(|(index, (model, field))| {
+            let is_duplicated = fields_to_be_added
+                .iter()
+                .filter(|(other_model, other_field)| model == other_model && field.name == other_field.name)
+                .count()
+                > 1;
+
+            if is_duplicated {
+                duplicated_relation_fields.push(index);
+            }
+        });
+
+    duplicated_relation_fields.iter().for_each(|index| {
+        let (_, ref mut field) = fields_to_be_added.get_mut(*index).unwrap();
+        let suffix = match &field.field_type {
+            FieldType::Relation(RelationInfo { name, .. }) => format!("_{}", &name),
+            FieldType::Base(_) => "".to_string(),
+            _ => "".to_string(),
+        };
+
+        field.name = format!("{}{}", field.name, suffix)
+    });
 }
 
 fn parse_int(value: &str) -> Option<i32> {
